refactor: character->avatar 변수명 변경
