+++ src/apis/avatar.ts
@@ -2,19 +2,19 @@ import { useMutation } from '@tanstack/react-query';
 
 export const domain = 'http://43.201.96.192:8080';
 
-export interface Character {
+export interface Avatar {
   id: number; // 고유값
-  characterId: number;
+  avatarId: number;
   name: string;
   rarity: string;
   dropRate: number;
 }
 
-export interface CharacterList {
-  avatars: Character[];
+export interface AvatarList {
+  avatars: Avatar[];
 }
 
-export const gachaCharacter = async (count: number) =>
+export const gachaAvatar = async (count: number) =>
   await fetch(`${domain}/gacha/draw`, {
     method: 'POST',
     headers: {
@@ -25,16 +25,16 @@ export const gachaCharacter = async (count: number) =>
     }),
   })
     .then(res => res.json())
-    .then(data => data as CharacterList);
+    .then(data => data as AvatarList);
 
-export const getCharacterList = async () =>
-  await fetch(`${domain}/character`, {
+export const getAvatarList = async () =>
+  await fetch(`${domain}/avatar`, {
     headers: {
       'Content-Type': 'application/json',
     },
   })
     .then(res => res.json())
-    .then(data => data as CharacterList);
+    .then(data => data as AvatarList);
 
 // 명세 미작성
 // export const saleCharacter = async (characterList: CharacterList) =>
@@ -48,9 +48,9 @@ export const getCharacterList = async () =>
 //     }),
 //   });
 
-export const useGachaCharacterApi = () => {
-  const { mutateAsync } = useMutation<CharacterList, Error, number>({
-    mutationFn: (count: number) => gachaCharacter(count),
+export const useGachaAvatarApi = () => {
+  const { mutateAsync } = useMutation<AvatarList, Error, number>({
+    mutationFn: (count: number) => gachaAvatar(count),
   });
 
   return mutateAsync;
